w08-d02, added method to filter tasks by activeness

+++ week-08/day-02/connecttomysql/src/main/java/com/greenfox/connecttomysql/repository/TodoRepo.java
@@ -3,5 +3,5 @@ package com.greenfox.connecttomysql.repository;
 import com.greenfox.connecttomysql.models.Todo;
 import org.springframework.data.repository.CrudRepository;
 
-public interface TodoRepo extends CrudRepository<Todo, Integer>{
+public interface TodoRepo extends CrudRepository<Todo, Long>{
 }
